src/async_modal/requirements.txt
aioredis==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
orjson==3.9.10
passlib==1.7.4
pendulum==3.0.0
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sse==1.2
sse-starlette==1.8.2
starlette==0.32.0.post1
time-machine==2.14.1
transliterate==1.10.2
typer==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
